requirements.txt
antlr4-python3-runtime==4.9.3
autograd==1.3
certifi==2021.10.8
cma==2.7.0
cycler==0.11.0
future==0.18.2
kiwisolver==1.3.2
matplotlib~=3.4.2
networkx==2.6.3
# numpy==1.24.0
pandas~=1.4.0
Pillow==8.4.0
progressbar2==3.55.0
pymoo==0.5.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv~=0.19.1
python-utils==2.5.6
peewee==3.14.4
pytz==2021.3
scipy~=1.7.3
six==1.16.0
wincertstore==0.2
scikit-learn~=1.0.2
nltk==3.6.5
gensim==4.1.2
pyspellchecker==0.6.2
pymop~=0.2.4
deap~=1.3.1
joblib~=1.0.1
imblearn~=0.0
# java8speedy~=1.2.0
setuptools~=58.2.0
seaborn~=0.11.2
tensorflow
